Create README.md
Many businesses are using social media networks to deliver different services and connect with clients and collect information about the thoughts and views of individuals. Sentiment analysis is a technique of machine learning that senses polarities such as positive or negative thoughts within the text, full documents, paragraphs, lines, or subsections. The aim of this study is to use context analysis to classify text in social media. People are freely expressing themselves on social networks, and millions of messages appear daily on popular websites for micro-blogging, such as Twitter, Tumblr, Facebook, etc. Therefore, these are the best places to begin to analyze people’s feelings. These analyses have potential widespread applicability to various fields, ranging from the detection of psychological disorders to profit and loss analyses of the market for various productions. The goal and primary objectives of this project was to analytically categorize and analyze the prevalent research techniques and implementations of Machine Learning techniques for Sentiment Analysis on various applications.